# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/checkout-session:
    post:
      operationId: createOrderCheckoutSession
      tags: [orders]
      summary: Create a Stripe Checkout session to pay for this order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"

  # DELIVERY TRACKING
...
components:
  schemas:
...
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Stripe-hosted Checkout page URL to redirect the browser to.

    # STORAGE
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, packed, shipped, out_for_delivery, delivered, cancelled, refunded]
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        shippingAddress:
          type: string
        shippingType:
          type: string
          enum: [global, pickup]
        shippingCarrier:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, items, subtotal, shippingCost, total, status, paymentMethod, paymentStatus, shippingAddress, shippingType, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [id, productId, productName, productImage, quantity, price, subtotal]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]
    OrderInput:
      type: object
      required: [items, shippingAddress, shippingType, paymentMethod]
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required: [productId, quantity]
        shippingAddress:
          type: string
        shippingType:
          type: string
          enum: [global, pickup]
        paymentMethod:
          type: string
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        paymentStatus:
          type: string
        trackingNumber:
          type: string
        shippingCarrier:
          type: string
        notes:
          type: string

    # DELIVERY TRACKING
...